HU 317: Visualización de clases del docente - actualizaciones
Mejora en los errores de carga y clases vacías.
Se agregó una opción para reintentar la consulta.
Se corrigieron dependencias de useFocusEffect.
Se conviritió el número de estudiantes a tipo entero.
Se ajustó la pluralización de estudiantes inscritos.

+++ backend/src/models/Clase.js
@@ -23,7 +23,7 @@ class Clase {
   static async listarPorDocente(idDocente) {
     const { rows } = await pool.query(
       `SELECT c.id, c.nombre, c."codigoUnico", c.curso, c."idDocente",
-              COUNT(ec."idEstudiante") AS "numEstudiantes"
+              COUNT(ec."idEstudiante")::INTEGER AS "numEstudiantes"
        FROM CLASE c
        LEFT JOIN ESTUDIANTE_CLASE ec ON ec."idClase" = c.id
        WHERE c."idDocente" = $1
